fix(configurator): integrate manual into configurator
